Fix building for 32 bit and add some hardening compiler flags

+++ include/ichor/Common.h
@@ -124,9 +124,9 @@ namespace Ichor {
         using key_equal = std::equal_to<>;  // Pred to use
         using hash_type = ankerl::unordered_dense::hash<std::string_view>;  // just a helper local type
 
-        size_t operator()(std::string_view txt) const   { return hash_type{}(txt); }
-        size_t operator()(const std::string& txt) const { return hash_type{}(txt); }
-        size_t operator()(const char* txt) const        { return hash_type{}(txt); }
+        uint64_t operator()(std::string_view txt) const   { return hash_type{}(txt); }
+        uint64_t operator()(const std::string& txt) const { return hash_type{}(txt); }
+        uint64_t operator()(const char* txt) const        { return hash_type{}(txt); }
     };
 
     template <
